// activity-07/frontend/src/types/company.ts
export interface CompanyShort {
  company_id: number;
  name: string;
  short_name: string;
  logo_url: string;
  category: string;
  operating_countries: string;
  office_locations: string;
  employee_size: string;
  yoy_growth_rate: string;
}